+++ src/ui/components/layout/status-bar/NotificationBanner.vue
